% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtphase1.R
\docType{package}
\name{dtphase1}
\alias{dtphase1}
\alias{dtphase1-package}
\title{dtphase1: A package that implements Phase 1 of the Decision Tree Framework}
\description{
This package implements Phase 1 as a shiny module. If you've never used
modules before, it might be helpful to review the \href{http://shiny.rstudio.com/articles/modules.html}{introduction on the RStudio website}. It's not strictly necessary though.
}
\details{
For usage examples, see ?\code{phase1UI} or ?\code{phase1.}
}

