% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{get_flowchart}
\alias{get_flowchart}
\title{Generate the phase 1 flowchart for the shiny app.}
\usage{
get_flowchart()
}
\description{
Make sure you update the flowchart.png that's used in the report as well...
}

