% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{phase1}
\alias{phase1}
\title{The server component of our phase 1 Shiny module.}
\usage{
phase1(input, output, session)
}
\arguments{
\item{input}{mandatory param by Shiny}

\item{output}{mandatory param by Shiny}

\item{session}{mandatory param by Shiny}
}
\description{
This function should be called using \code{callModule} in the server function
of your shiny app. It handles the downloading of the report.
}
\details{
In callModule, use the template \code{callModule(phase1, id)}, where id is
the same string that you use in phase1UI.
}
\examples{
\dontrun{
# minimal example
ui <- phase1UI("p1")
server <- function(input, output, session) {
 callModule(phase1, "p1")
}
}

}

