% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{phase1UI}
\alias{phase1UI}
\title{Create the UI for Phase 1.}
\usage{
phase1UI(id)
}
\arguments{
\item{id}{the ID that will be used to uniquely identify all of the ui
components in the shiny module. Must be the same as the id used in \code{callModule}.}
}
\value{
UI components that collectively represent the climate screening worksheet of Phase 1.
}
\description{
This creates all of the UI for our Phase 1 climate screening worksheet. You
should call it in the ui function of your shiny app.
}
\examples{
\dontrun{
# minimal example
ui <- fluidPage(
  phase1UI("p1")
)
server <- function(input, output, session) {
 callModule(phase1, "p1")
}
}
}

